+++ src/main/java/com/example/demo/controller/FoundController.java
@@ -3,8 +3,13 @@ package com.example.demo.controller;
 import com.example.demo.domain.Found;
 import com.example.demo.dto.FoundDto;
 import com.example.demo.service.FoundService;
+import com.example.demo.vo.AuthUserVo_j;
+import com.example.demo.vo.FoundVO;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -17,17 +22,43 @@ public class FoundController {
         foundService.saveFound(found);
         return found;
     }
-
+    @GetMapping("")
+    FoundDto getFoundDto(@PathVariable int foundId) {
+        FoundDto foundDto= foundService.getFoundDto(foundId);
+        return foundDto;
+    }
     @PostMapping("/{foundId}/edit")
     public FoundDto updateFound(@PathVariable int foundId, @RequestBody FoundDto foundDto) {
-        foundService.updateFound(foundId, foundDto);
-        return foundDto;
+        FoundDto check = getFoundDto(foundId);
+        AuthUserVo_j auth = (AuthUserVo_j) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int id = auth.getId();
+        if(id == check.getUserId()){
+            foundService.updateFound(foundId, foundDto);
+            return foundDto;
+        }
+        else {
+            System.out.println("Invalid User");
+            return null;
+        }
     }
 
     @DeleteMapping("{foundId}/delete")
     public void deleteFound(@PathVariable int foundId){
-        foundService.deleteFound(foundId);
+
+        FoundDto check = getFoundDto(foundId);
+        AuthUserVo_j auth = (AuthUserVo_j) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int id = auth.getId();
+        if(id == check.getUserId()){
+            foundService.deleteFound(foundId);
+        }
+        else {
+            System.out.println("Invalid User");
+        }
     }
-    
+    @GetMapping("/list")
+    public List<FoundVO> getList(){
+        List<FoundVO> list = foundService.getList();
+        return list;
+    }
 
 }
